dentist_website/requirements.txt
aiofiles==0.5.0
asgiref==3.4.1
asttokens==2.0.5
blessed==1.19.0
cached-property==1.5.2
certifi==2021.10.8
colorama==0.4.4
Command==0.1.0
commonmark==0.9.1
diot==0.1.5
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.10
django-extensions==3.1.5
django-on-heroku==1.1.2
django-static-jquery==2.1.4
enlighten==1.10.2
executing==0.8.2
filelock==3.4.2
gunicorn==20.1.0
inflection==0.5.1
Jinja2==3.0.3
liquidpy==0.7.3
Markdown==3.3.6
MarkupSafe==2.0.1
more-itertools==8.12.0
numpy==1.22.0
pandas==1.3.5
Pillow==9.0.1
pipda==0.4.5
pipen==0.2.13
pipenv==2021.11.23
platformdirs==2.4.1
prefixed==0.3.2
psutil==5.9.0
psycopg2-binary==2.9.3
pure-eval==0.2.1
Pygments==2.11.2
pyparam==0.4.16
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python-frontmatter==1.0.0
python-simpleconf==0.4.2
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
rich==10.16.2
simplug==0.0.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
uvloop==0.16.0
varname==0.8.1
virtualenv==20.13.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
whitenoise==6.1.0
xqute==0.0.6
